Remove commands and map to a generic text command
We were creating separate commands for each of the various specific
commands that we were adding. This turned out to be a bad idea, as
1/ it's a lot of repetetive junk and 2/ it will eventually make the
notion of aliases a bit hard to implement.

We now have a generic text command. We can add specific commands, but
anything a user enters as raw text should be passed along using this
command.