WebSockets: Handle connection disconnects a bit
Currently, there's no indication when a connection to the server cannot be
established. This patch does a _mildly_ better job in handling this by
attempting reconnects every 5 seconds. This works pretty well, but doesn't
work super great (for some reason) when the react server restarts itself.

Teaching sand to think was a mistake.